Rhino DOES NOT preserve the property creation order of "negative integer" since it seems to be incorrectly treating these as an integer index even though they're not in the defined range of 0 <= i <= F (https://262.ecma-international.org/13.0/#sec-object-type)

+++ src/org/mozilla/javascript/json/JsonParser.java
@@ -118,11 +118,12 @@ public class JsonParser {
                     consume(':');
                     value = readValue();
 
-                    long index = ScriptRuntime.indexFromString(id);
+                    //HtmlUnit
+                    int index = (int) ScriptRuntime.indexFromString(id);
                     if (index < 0) {
                         object.put(id, object, value);
                     } else {
-                        object.put((int) index, object, value);
+                        object.put(index, object, value);
                     }
 
                     needsComma = true;
